added msg

+++ main.py
@@ -340,14 +340,18 @@ async def test(ctx):
         if "VNTA" in role.name:
             print(role.id, role.name)
 
+bot.pendings = {}
 @bot.command()
 async def link(ctx, *, ign):
     if not any(allow in [role.id for role in ctx.author.roles] for allow in accepted):
         return await ctx.reply("Only VNTA members are given the exclusive rights to use the bot.")
-    bot.links[str(ctx.author.id)] = str(ign)
-    embed = discord.Embed(description=f"✅ Linked successfully!", color=5963593)
+    staff = bot.get_channel(813447381752348723)
+    a = await staff.send(f"{ctx.author.mention} wants to link with {ign}")
+    await a.add_reaction("✅")
+    await a.add_reaction("❌")
+    bot.pendings[a.id] = {ctx.author.id:str(ign)}
+    embed = discord.Embed(description=f"Link request submitted to staff successfully!", color=5963593)
     await ctx.reply(embed=embed)
-    await update_links()
 
 @bot.command(aliases=["con"])
 async def contract(ctx, *, ign=None):
@@ -434,11 +438,22 @@ async def on_connect():
 
 @bot.event
 async def on_raw_reaction_add(payload):
-    if payload.user_id != 771601176155783198:
-        return
-    if str(payload.emoji) == "🗑️":
+    if str(payload.emoji) == "🗑️" and payload.user_id == 771601176155783198:
         chl = await bot.fetch_channel(payload.channel_id)
         msg = await chl.fetch_message(payload.message_id)
         await msg.delete()
+    if payload.channel_id == 813447381752348723 and str(payload.emoji) in ["✅", "❌"]:
+        if str(payload.emoji) == "❌":
+            userd = bot.pendings[payload.message_id]
+            user = userd.keys()[0]
+            user = bot.get_user(user)
+            await user.send(f"❌ Your request to link with `{userd.values()[0]}` is denied!")
+        else:
+            userd = bot.pendings[payload.message_id]
+            user = userd.keys()[0]
+            user = bot.get_user(user)
+            bot.links[str(user.id)] = userd.values()[0]
+            await update_links()
+            await user.send(f"✅ Your request to link with `{userd.values()[0]}` is denied!")
 
 bot.run("ODUzOTcxMjIzNjgyNDgyMjI2.YMdIrQ.N-06PP7nmUz-E-3bQvWqCtArhP0")
